Move aelo plotting functions to a common file to be imported from plots.py and from compute_rtgm.py
