+++ openquake/calculators/postproc/compute_rtgm.py
@@ -52,7 +52,7 @@ from openquake.baselib import hdf5
 from openquake.hazardlib.imt import from_string
 from openquake.hazardlib.calc.mean_rates import to_rates
 from openquake.calculators import postproc
-from openquake.commands.plot import (
+from openquake.calculators.postproc.aelo_plots import (
     plot_mean_hcurves_afe_rtgm, plot_disagg_by_src, plot_governing_mce,
     _find_fact_maxC)
